ccap: Add support for direct image output onto frame buffer
Introduced a new feature in the CCAP driver to allow captured images to be
output directly onto the frame buffer using ioctl control.
This eliminates the overhead of image conversion and transfer.
Also added support for rendering captured images in their original size
at specified positions on the frame buffer.

Change-Id: Iddceaee47b6478d74f1f06ab677e539e760334df